+++ src/main.rs
@@ -9,6 +9,7 @@ mod day_6;
 mod day_7;
 mod day_8;
 mod day_9;
+mod util;
 
 use std::io::{self, Write};
 use std::time::Instant;
